refactor: relocate config
