+++ cmd/notification-service/wire_gen.go
@@ -10,7 +10,7 @@ import (
 	"context"
 	"github.com/ThreeDotsLabs/watermill/message"
 	"github.com/openmeterio/openmeter/app/common"
-	"github.com/openmeterio/openmeter/config"
+	"github.com/openmeterio/openmeter/app/config"
 	"github.com/openmeterio/openmeter/openmeter/ent/db"
 	"github.com/openmeterio/openmeter/openmeter/meter"
 	"github.com/openmeterio/openmeter/openmeter/streaming"
