+++ openmeter/watermill/driver/kafka/broker.go
@@ -10,7 +10,7 @@ import (
 	"github.com/IBM/sarama"
 	otelmetric "go.opentelemetry.io/otel/metric"
 
-	"github.com/openmeterio/openmeter/config"
+	"github.com/openmeterio/openmeter/app/config"
 	"github.com/openmeterio/openmeter/openmeter/watermill/driver/kafka/metrics"
 )
